Architecture Changes
changes:
add collidable class to keep track of what can be collided with in the
game
Change killableInGame to a static class - only need one instance
changed from bullet.hit() to every other element incorperating a
hitByBullet() method...bullet cant keep track of everything and what to
do
when hitting specific things but any specific thing can keep track of
that for itself
bullets now have a property called damage
bullets no longer have a hit() method, but a getDamage method which
returns its damage

KNOWN ISSUES:
Removing bullets doesnt seem to be working (this involves the
BasicAmmo.hitbyBullet method and the Collidable.removeCollidable
methods)

+++ Bosem/scripts/KillableInGame.js
@@ -1,44 +1,58 @@
 ﻿var Bosem;
 (function (Bosem) {
     var KillableInGame = (function () {
-        function KillableInGame(game, numPlayers) {
+        function KillableInGame() {
+        }
+        KillableInGame.init = function (game, numPlayers) {
             this.game = game;
             this.players = [];
             this.teams = [[numPlayers], [1]];
             for (var i = 0; i < numPlayers; i++) {
                 this.players.push(new Bosem.Player(this.game, this.game.width - this.game.width / (i + 1), 30, i));
                 this.teams[i].push(this.players[i]);
+                Bosem.Collidable.addCollidable(this.players[i]);
             }
 
             //need to get rid of this eventually, make it an array of enemies (or no enemies at all...just deal with allies) but for now, dont want to deal
             this.players[0].setEnemy(this.players[1]);
             this.players[1].setEnemy(this.players[0]);
-        }
-        KillableInGame.prototype.addKillable = function (killable, teamNum) {
-            this.teams[teamNum].push(killable);
         };
-        KillableInGame.prototype.getAllies = function (teamNum) {
+        KillableInGame.addKillable = function (killable, teamNum) {
+            this.teams[teamNum].push(killable);
+            Bosem.Collidable.addCollidable(killable);
+        };
+        KillableInGame.getAllies = function (teamNum) {
             return this.teams[teamNum];
         };
-        KillableInGame.prototype.getPlayers = function () {
+        KillableInGame.getPlayers = function () {
             return this.players;
         };
-        KillableInGame.prototype.killPlayer = function (player) {
-            player.lives--;
-            var killed = player;
-            player.kill();
+        KillableInGame.getPlayersGroup = function () {
+            var group = new Phaser.Group(this.game);
+            for (var i = 0; i < this.players.length; i++) {
+                group.add(this.players[i]);
+            }
+            return group;
+        };
+        KillableInGame.killPlayer = function (player) {
+            if (player.canDie) {
+                player.canDie = false;
+                player.lives--;
+                var killed = player;
+                player.kill();
 
-            setTimeout(function () {
-                killed.revive();
-                killed.canDie = true;
-                killed.position.set(Math.floor(Math.random() * killed.game.world.width), 100);
-                killed.checkWorldBounds = true;
-                killed.hp = 1000;
-            }, 100);
+                setTimeout(function () {
+                    killed.revive();
+                    killed.canDie = true;
+                    killed.position.set(Math.floor(Math.random() * killed.game.world.width), 100);
+                    killed.checkWorldBounds = true;
+                    killed.canDie = true;
+                }, 5000);
+            }
         };
 
         //checks to be made
-        KillableInGame.prototype.update = function () {
+        KillableInGame.update = function () {
             for (var i = 0; i < this.teams.length; i++) {
                 for (var j = 0; j < this.teams[i].length; j++) {
                     if (this.teams[i][j] instanceof Bosem.Player && this.teams[i][j].hp <= 0) {
@@ -50,7 +64,7 @@
             }
         };
 
-        KillableInGame.prototype.killKillable = function (killable) {
+        KillableInGame.killKillable = function (killable) {
             //implement
         };
         return KillableInGame;
