+++ Bosem/scripts/KillableInGame.ts
@@ -2,47 +2,62 @@
     export class KillableInGame {
 
         //adds players in the beginnign of the game
-        players: Player[]; //players in game
-        teams: any[][]; //team of any player
-        game: Phaser.Game;
-        constructor(game: Phaser.Game, numPlayers: number) {
+       static players: Player[]; //players in game
+       static teams: any[][]; //team of any player
+        static game: Phaser.Game;
+        
+        static init(game:Phaser.Game, numPlayers :number) {
             this.game = game;
             this.players = [];
             this.teams = [[numPlayers], [1]];
             for (var i = 0; i < numPlayers; i++) {
                 this.players.push(new Player(this.game, this.game.width - this.game.width / (i + 1), 30, i));
                 this.teams[i].push(this.players[i]);
+                Collidable.addCollidable(this.players[i]);
             }
+
 
             //need to get rid of this eventually, make it an array of enemies (or no enemies at all...just deal with allies) but for now, dont want to deal
             this.players[0].setEnemy(this.players[1]);
             this.players[1].setEnemy(this.players[0]);
         }
-        addKillable(killable:any, teamNum:number) {
-            this.teams[teamNum].push(killable);
+       static addKillable(killable:any, teamNum:number) {
+           this.teams[teamNum].push(killable);
+           Collidable.addCollidable(killable);
         }
-        getAllies(teamNum: number) {
+        static getAllies(teamNum: number) {
             return this.teams[teamNum];
         }
-        getPlayers() {
+        static getPlayers() {
             return this.players;
         }
-        killPlayer(player: Player) {
-            player.lives--;
-            var killed = player;
-            player.kill();
+        static getPlayersGroup() {
+            var group: Phaser.Group = new Phaser.Group(this.game);
+            for(var i = 0; i < this.players.length; i++){
+                group.add(this.players[i]);
+             }
+            return group;
+        }
+        static  killPlayer(player: Player) {
+            if (player.canDie) {
+                player.canDie = false;
+                player.lives--;
+                var killed = player;
+                player.kill();
 
-            setTimeout(function () {
-                killed.revive();
-                killed.canDie = true;
-                killed.position.set(Math.floor(Math.random() * killed.game.world.width), 100);
-                killed.checkWorldBounds = true;
-                killed.hp = 1000;
-            }, 100);
+                setTimeout(function () {
+                    killed.revive();
+                    killed.canDie = true;
+                    killed.position.set(Math.floor(Math.random() * killed.game.world.width), 100);
+                    killed.checkWorldBounds = true;  
+                    killed.canDie = true;
+
+                }, 5000);
+            }
         }
 
         //checks to be made
-        update() {
+        static  update() {
             for (var i = 0; i < this.teams.length; i++) {
                 for (var j = 0; j < this.teams[i].length; j++) {
                     if (this.teams[i][j] instanceof Player && this.teams[i][j].hp <=0) {
@@ -55,7 +70,7 @@
             }
         }
 
-        killKillable(killable: any) {
+        static   killKillable(killable: any) {
             //implement
         }
